requirements.txt
langchain
openai
faiss-cpu
streamlit
python-dotenv
PyPDF2
